fix: fix circular build
