+++ packages/connection/src/common/connection/drivers/index.ts
@@ -1,5 +1,4 @@
 export * from './base';
-export * from './utils';
 export * from './node-message-port';
 export * from './socket';
 export * from './ws-websocket';
